Adicionando autenticacao via JWT e filtros nas entidades

+++ src/main/java/com/gft/loja/domain/service/ProdutoService.java
@@ -25,12 +25,33 @@ public class ProdutoService {
 	@Autowired
 	private EstoqueRepository estoqueRepository;
 
-	public List<Produto> listar() {
+	@Autowired
+	private EstoqueService estoqueService;
+
+	public List<Produto> listarTodos() {
 		return produtoRepository.findAll();
 	}
+	
+	public List<Estoque> listar() {
+		return estoqueRepository.findByQuantidadeGreaterThanAndValorVendaIsNotNull(0);
+	}
+	
+	public List<Estoque> listarAsc() {
+		return estoqueRepository.findByQuantidadeGreaterThanAndValorVendaIsNotNullOrderByProdutoDescricaoAsc(0);
+	}
+	
+	public List<Estoque> listarDesc() {
+		return estoqueRepository.findByQuantidadeGreaterThanAndValorVendaIsNotNullOrderByProdutoDescricaoDesc(0);
+	}
+	
+	public List<Estoque> listarComFiltro(String produtoDesc) {
+		return estoqueRepository.findByQuantidadeGreaterThanAndValorVendaIsNotNullAndProdutoDescricaoContaining(0,produtoDesc);
+	}
+	
 
 	public Produto buscar(Long id) {
-		return produtoRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Produto não encontrado. Faça o preenchimento correto e tente novamente"));
+		return produtoRepository.findById(id).orElseThrow(() -> new NoSuchElementException(
+				"Produto não encontrado. Faça o preenchimento correto e tente novamente"));
 
 	}
 
@@ -45,18 +66,24 @@ public class ProdutoService {
 
 		Produto produtoSalvo = produtoRepository.save(produto);
 		Estoque estoque = new Estoque(produtoSalvo.getId(), produto, 0, null); // No momento que cria o
-																							// produto, cria um estoque
-																							// com quantidade e valor 0
+																				// produto, cria um estoque
+																				// com quantidade e valor 0
 		estoqueRepository.save(estoque);
 
 		return produtoSalvo;
 	}
-	
-	
 
 	public void excluir(Long produtoId) {
 		try {
-			produtoRepository.deleteById(produtoId);
+			Estoque produtoEstoque = estoqueService.buscar(produtoId);
+			if (produtoEstoque.getQuantidade() == 0) {
+
+				estoqueRepository.deleteById(produtoId);
+				produtoRepository.deleteById(produtoId);
+			} else {
+				throw new EntidadeEmUsoException(
+						"O produto possui uma quantidade no estoque. Favor excluir o produto do estoque para prosseguir.");
+			}
 
 		} catch (DataIntegrityViolationException e) {
 			throw new EntidadeEmUsoException("Produto está em uso por outra entidade.");
@@ -66,10 +93,14 @@ public class ProdutoService {
 
 	}
 
-	public Produto atualizar(Long produtoId,Produto produto) {
+	public Produto atualizar(Long produtoId, Produto produto) {
 		Produto produtoBuscado = buscar(produtoId);
-		BeanUtils.copyProperties(produto, produtoBuscado,"id");
+		BeanUtils.copyProperties(produto, produtoBuscado, "id");
 		salvar(produtoBuscado);
 		return produtoBuscado;
 	}
+
+	
+
+	
 }
